+++ gazebo/gui/model/ModelEditor.hh
@@ -40,12 +40,15 @@ namespace gazebo
       /// \brief Destuctor.
       public: virtual ~ModelEditor();
 
-      /// \brief QT callback when entering building edit mode
+      /// \brief QT callback when entering model edit mode
       /// \param[in] _checked True if the menu item is checked
       private slots: void OnEdit(bool _checked);
 
       /// \brief Callback when the model has been completed.
       private: void OnFinish();
+
+      /// \brief Toggle main window's toolbar to display model editor icons.
+      private: void ToggleToolbar();
 
       /// \brief Contains all the model editor tools.
       private: ModelEditorPalette *modelPalette;